[#163386375] separate preferences to be persisted from those that must not
