+++ ts/store/actions/preferences.ts
@@ -4,11 +4,6 @@ export const preferencesLanguagesLoadSuccess = createStandardAction(
   "PREFERENCES_LANGUAGES_LOAD_SUCCESS"
 )<ReadonlyArray<string>>();
 
-export const preferenceFingerprintIsEnabledSave = createStandardAction(
-  "PREFERENCES_FINGERPRINT_SAVE_SUCCESS"
-)<boolean>();
-
 export type PreferencesActions = ActionType<
-  | typeof preferenceFingerprintIsEnabledSave
-  | typeof preferencesLanguagesLoadSuccess
+  typeof preferencesLanguagesLoadSuccess
 >;
