+++ ts/store/reducers/preferences.ts
@@ -1,17 +1,14 @@
 import { isActionOf } from "typesafe-actions";
 
-import { preferenceFingerprintIsEnabledSave } from "../actions/preferences";
 import { preferencesLanguagesLoadSuccess } from "../actions/preferences";
 import { Action } from "../actions/types";
 
 export type PreferencesState = Readonly<{
   languages: ReadonlyArray<string> | undefined;
-  isFingerprintEnabled: boolean | undefined;
 }>;
 
 const initialPreferencesState: PreferencesState = {
-  languages: undefined,
-  isFingerprintEnabled: undefined
+  languages: undefined
 };
 
 export default function preferencesReducer(
@@ -22,12 +19,6 @@ export default function preferencesReducer(
     return {
       ...state,
       languages: action.payload
-    };
-  }
-  if (isActionOf(preferenceFingerprintIsEnabledSave, action)) {
-    return {
-      ...state,
-      isFingerprintEnabled: action.payload
     };
   }
